+++ CHANGELOG.md
@@ -8,3 +8,6 @@ v0.2.0
   first element in a tag is not a string.
 - Fixed input structures being modified by AntiHTML.
 - Added location information to errors reported in nested structures.
+- Added support for specifying Comment and Text nodes at the top level
+  to htmlDocument and htmlFragment.
+- Added support for specifying multiple top level tags to htmlFragment.